Simplify Bidirectional Link Logic
The current implementation relies heavily on
ModelManager.java to update the list in
Contact when linking a contact to an event.

Let's simplify the process by moving the bulk of
the logic to EventLinkCommand.java.

This ensures that linking is handled by
EventLinkCommand.java, separating areas
of concern for the classes.